cleaned up file naming/breakdown
