+++ internal/core/workflow_archive_connector.go
@@ -1,0 +1,5 @@
+package core
+
+const (
+	WorkflowNameArchiveConnectorV1               = "core.connector.archive.v1"
+)
